Moved strategies to strategies package
