+++ src/org/usfirst/frc4904/autonomous/strategies/OpenPortcullis.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.autonomous.commands;
+package org.usfirst.frc4904.autonomous.strategies;
 
 
 import org.usfirst.frc4904.standard.subsystems.chassis.Chassis;
