+++ src/org/usfirst/frc4904/robot/Robot.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc4904.robot;
 
 
-import org.usfirst.frc4904.autonomous.commands.CrossLowbar;
+import org.usfirst.frc4904.autonomous.strategies.CrossLowbar;
 import org.usfirst.frc4904.robot.humaninterface.drivers.Nathan;
 import org.usfirst.frc4904.robot.humaninterface.drivers.NathanGain;
 import org.usfirst.frc4904.robot.humaninterface.operators.DefaultOperator;
